refactor(graphql): split GraphQL resolvers to files for maintainability
done:
- graphql/src/resolvers/getResolvers.ts now returns all used resolvers
- graphql/src/resolvers/ and graphql/src/resolvers/query/ have now
  >10 resolvers total to better organize the resolvers, and make
  graphql/src/index.ts file smaller and more maintainable

also:
 - move types to graphql/src/types.ts etc, also partially to avoid
   circular dependencies

refs US-171
