+++ graphql/src/index.ts
@@ -3,7 +3,6 @@ import './sentry-init.mjs'; // MUST be FIRST IMPORT! See file for details.
 
 import http from 'http';
 
-import type { GraphQLResolveInfoWithCacheControl } from '@apollo/cache-control-types';
 import { ApolloServer } from '@apollo/server';
 import { ApolloServerPluginLandingPageDisabled } from '@apollo/server/plugin/disabled';
 import { ApolloServerPluginDrainHttpServer } from '@apollo/server/plugin/drainHttpServer';
@@ -22,44 +21,11 @@ import {
   cspConfigWithoutApolloSandbox,
 } from './constants.js';
 import { ElasticSearchAPI } from './datasources/es/index.js';
-import type {
-  AccessibilityProfileType,
-  AdministrativeDivisionParams,
-  ElasticSearchIndex,
-  OntologyTreeParams,
-  OntologyWordParams,
-  OrderByDistanceParams,
-  OrderByNameParams,
-  SuggestionsParams,
-} from './datasources/es/types.js';
 import healthz from './healthz.js';
 import readiness from './readiness.js';
-import pageInfoResolver, { Edge } from './resolvers/pageInfoResolver.js';
-import { elasticSearchSchema } from './schemas/es.js';
-import { geoSchema } from './schemas/geojson.js';
-import { locationSchema } from './schemas/location.js';
-import { ontologySchema } from './schemas/ontology.js';
-import { palvelukarttaSchema } from './schemas/palvelukartta.js';
-import { querySchema } from './schemas/query.js';
-import { sharedSchema } from './schemas/shared.js';
-import type {
-  ConnectionArguments,
-  ConnectionCursorObject,
-  EsHit,
-  EsResults,
-  Venue,
-} from './types.js';
-import {
-  createCursor,
-  edgesFromEsResults,
-  elasticLanguageFromGraphqlLanguage,
-  findClosestEnvFile,
-  getEsOffsetPaginationQuery,
-  getHits,
-  getTodayString,
-  sentryConfig,
-  validateOrderByArguments,
-} from './utils.js';
+import { getResolvers } from './resolvers/getResolvers.js';
+import { getSchemas } from './schemas/getSchemas.js';
+import { findClosestEnvFile, sentryConfig } from './utils.js';
 
 const ENV_FILE_PATH = findClosestEnvFile();
 
@@ -85,323 +51,14 @@ function getLandingPagePlugin() {
   return ApolloServerPluginLandingPageDisabled();
 }
 
-function getCspConfig():
-  | typeof cspConfigWithApolloSandbox
-  | typeof cspConfigWithoutApolloSandbox {
-  if (ENABLE_APOLLO_SANDBOX) {
-    return cspConfigWithApolloSandbox;
-  }
-  return cspConfigWithoutApolloSandbox;
-}
-
-type UnifiedSearchQuery = {
-  text?: string;
-  ontology?: string;
-  administrativeDivisionIds?: string[];
-  ontologyTreeIdOrSets?: string[][];
-  ontologyWordIdOrSets?: string[][];
-  providerTypes?: string[];
-  serviceOwnerTypes?: string[];
-  targetGroups?: string[];
-  mustHaveReservableResource?: boolean;
-  index?: ElasticSearchIndex;
-  languages?: string[];
-  openAt?: string;
-  orderByDistance?: OrderByDistanceParams;
-  orderByName?: OrderByNameParams;
-  orderByAccessibilityProfile?: AccessibilityProfileType;
-} & ConnectionArguments;
-
-type VenueProps = {
-  venue: Venue;
-};
-
-type QueryContext = {
-  dataSources: {
-    elasticSearchAPI: ElasticSearchAPI;
-  };
-};
-
-type GeoJSONPointInput = {
-  geometry?: {
-    coordinates?: {
-      longitude?: number;
-      latitude?: number;
-    };
-  };
-  longitude?: number;
-  latitude?: number;
-};
-
-type LongitudeLatitude = [longitude: number, latitude: number];
-
-type PageInfoProps = {
-  edges: Edge[];
-  hits: number;
-  connectionArguments: ConnectionArguments;
-};
-
-const resolvers = {
-  Query: {
-    unifiedSearch: async (
-      _source: unknown,
-      {
-        text,
-        ontology,
-        administrativeDivisionIds,
-        ontologyTreeIdOrSets,
-        ontologyWordIdOrSets,
-        providerTypes,
-        serviceOwnerTypes,
-        targetGroups,
-        mustHaveReservableResource,
-        index,
-        after,
-        first,
-        languages,
-        openAt,
-        orderByDistance,
-        orderByName,
-        orderByAccessibilityProfile,
-      }: UnifiedSearchQuery,
-      { dataSources }: QueryContext,
-      info: GraphQLResolveInfoWithCacheControl
-    ) => {
-      const connectionArguments = { after, first };
-      const { from, size } = getEsOffsetPaginationQuery(connectionArguments);
-
-      validateOrderByArguments(
-        orderByDistance,
-        orderByName,
-        orderByAccessibilityProfile
-      );
-
-      const result = await dataSources.elasticSearchAPI.getQueryResults({
-        text,
-        ontology,
-        administrativeDivisionIds,
-        ontologyTreeIdOrSets,
-        ontologyWordIdOrSets,
-        providerTypes,
-        serviceOwnerTypes,
-        targetGroups,
-        mustHaveReservableResource,
-        index,
-        from,
-        size,
-        languages: elasticLanguageFromGraphqlLanguage(languages),
-        openAt,
-        orderByDistance,
-        orderByName,
-        orderByAccessibilityProfile,
-      });
-
-      const getCursor = (offset: number) =>
-        createCursor<ConnectionCursorObject>({
-          offset: from + offset,
-        });
-
-      // Find shared data
-      const edges = edgesFromEsResults(result, getCursor);
-      const hits = getHits(result);
-
-      if (result.hits.hits.length >= 1000) {
-        info.cacheControl.setCacheHint({
-          maxAge: parseInt(process.env.CACHE_MAX_AGE ?? '3600', 10),
-        });
-      }
-
-      return { es_results: [result], edges, hits, connectionArguments };
-    },
-    unifiedSearchCompletionSuggestions: async (
-      _source: unknown,
-      { prefix, languages, index, size }: SuggestionsParams,
-      { dataSources }: QueryContext
-    ) => {
-      const res = await dataSources.elasticSearchAPI.getSuggestions({
-        prefix,
-        languages: elasticLanguageFromGraphqlLanguage(languages),
-        size,
-        index,
-      });
-
-      return {
-        suggestions: res.suggest.suggestions[0].options.map((option) => ({
-          label: option.text,
-        })),
-      };
-    },
-    administrativeDivisions: async (
-      _,
-      args: AdministrativeDivisionParams,
-      { dataSources }: QueryContext
-    ) => {
-      const res: EsResults =
-        await dataSources.elasticSearchAPI.getAdministrativeDivisions(args);
-      return res.hits.hits.map((hit: EsHit) => ({
-        id: hit._id,
-        ...hit._source,
-      }));
-    },
-    ontologyTree: async (
-      _,
-      args: OntologyTreeParams,
-      { dataSources }: QueryContext
-    ) => {
-      const res: EsResults =
-        await dataSources.elasticSearchAPI.getOntologyTree(args);
-      return res.hits.hits.map((hit: EsHit) => ({
-        id: hit._id,
-        ...hit._source,
-      }));
-    },
-    ontologyWords: async (
-      _,
-      args: OntologyWordParams,
-      { dataSources }: QueryContext
-    ) => {
-      const res: EsResults =
-        await dataSources.elasticSearchAPI.getOntologyWords(args);
-
-      return res.hits.hits.map((hit: EsHit) => ({
-        id: hit._id,
-        ...hit._source,
-        label: hit._source.name,
-      }));
-    },
-  },
-
-  SearchResultConnection: {
-    count({ hits }: PageInfoProps) {
-      return hits;
-    },
-    max_score({ es_results }: unknown) {
-      return es_results[0].hits.max_score;
-    },
-    async pageInfo({ edges, hits, connectionArguments }: PageInfoProps) {
-      return await pageInfoResolver(edges, hits, connectionArguments);
-    },
-    edges({ edges }: PageInfoProps) {
-      return edges;
-    },
-    es_results({ es_results }: unknown) {
-      return es_results;
-    },
-  },
-
-  UnifiedSearchVenue: {
-    name({ venue }: VenueProps) {
-      return venue.name;
-    },
-    description({ venue }: VenueProps) {
-      return venue.description;
-    },
-    location({ venue }: VenueProps) {
-      return venue.location;
-    },
-    openingHours({ venue }: VenueProps) {
-      return venue.openingHours;
-    },
-    images({ venue }: VenueProps) {
-      return venue.images;
-    },
-    ontologyWords({ venue }: VenueProps) {
-      return venue.ontologyWords;
-    },
-    serviceOwner({ venue }: VenueProps) {
-      return venue.serviceOwner;
-    },
-    reservation({ venue }: VenueProps) {
-      return venue.reservation;
-    },
-    targetGroups({ venue }: VenueProps) {
-      return venue.targetGroups;
-    },
-    accessibility({ venue }: VenueProps) {
-      return venue.accessibility;
-    },
-    accessibilityShortcomingFor(
-      { venue }: VenueProps,
-      args: { profile?: AccessibilityProfileType }
-    ) {
-      if (args.profile) {
-        const shortcoming = venue.accessibility.shortcomings.filter(
-          (shortcoming) => shortcoming.profile === args.profile
-        );
-        if (shortcoming && shortcoming.length > 0) {
-          return shortcoming[0];
-        }
-      }
-      return undefined;
-    },
-    meta({ venue }: VenueProps) {
-      return venue.meta;
-    },
-  },
-
-  OpeningHours: {
-    today({ data }: unknown) {
-      const openingHoursToday = data.find(
-        (openingHoursDay) => openingHoursDay.date === getTodayString()
-      );
-      return openingHoursToday ? openingHoursToday.times : [];
-    },
-  },
-
-  RawJSON: {
-    data(parent: unknown) {
-      // Testing and debugging only
-      return JSON.stringify(parent);
-    },
-  },
-
-  GeoJSONCRSProperties: {
-    __resolveType() {
-      return null;
-    },
-  },
-  GeoJSONGeometryInterface: {
-    __resolveType() {
-      return 'GeoJSONPoint';
-    },
-  },
-  GeoJSONInterface: {
-    __resolveType() {
-      return null;
-    },
-  },
-  GeoJSONPoint: {
-    type() {
-      return 'Point';
-    },
-    coordinates(obj: GeoJSONPointInput): LongitudeLatitude | null {
-      const long = obj.geometry?.coordinates?.longitude ?? obj.longitude;
-      const lat = obj.geometry?.coordinates?.latitude ?? obj.latitude;
-
-      return long && lat ? [long, lat] : null;
-    },
-  },
-  GeoJSONFeature: {
-    type() {
-      return 'Point';
-    },
-    geometry(parent: unknown) {
-      return parent;
-    },
-  },
-};
+const getCspConfig = () =>
+  ENABLE_APOLLO_SANDBOX
+    ? cspConfigWithApolloSandbox
+    : cspConfigWithoutApolloSandbox;
 
 const combinedSchema = buildSubgraphSchema({
-  typeDefs: [
-    querySchema,
-    elasticSearchSchema,
-    palvelukarttaSchema,
-    locationSchema,
-    sharedSchema, // FIXME: Overlapping with events-proxy (https://tapahtumat-proxy.test.kuva.hel.ninja/proxy/graphql); Location Image
-    geoSchema,
-    ontologySchema,
-  ],
-  resolvers,
+  typeDefs: getSchemas(),
+  resolvers: getResolvers(),
 });
 
 void (async () => {
