+++ graphql/src/resolvers/OpeningHoursResolver.ts
@@ -1,0 +1,12 @@
+import { getTodayString } from '../utils.js';
+
+export function OpeningHoursResolver() {
+  return {
+    today({ data }: unknown) {
+      const openingHoursToday = data.find(
+        (openingHoursDay) => openingHoursDay.date === getTodayString()
+      );
+      return openingHoursToday ? openingHoursToday.times : [];
+    },
+  };
+}
