+++ graphql/src/resolvers/RawJSONResolver.ts
@@ -1,0 +1,8 @@
+export function RawJSONResolver() {
+  return {
+    data(parent: unknown) {
+      // Testing and debugging only
+      return JSON.stringify(parent);
+    },
+  };
+}
